chore: consolidate shared methods
