+++ setup.cfg
@@ -14,7 +14,7 @@ warn_unused_configs = True
 check_untyped_defs = True
 disallow_untyped_defs = True
 [tool:pytest]
-addopts = -p no:playwright -p no:asyncio --runpytest subprocess -vv
+addopts = -p no:playwright --runpytest subprocess -vv
 testpaths =
     tests
 [coverage:run]
